[손희진] saveManager 수정 저장/로드 테스트.

+++ Assets/Scripts/Contents/RecipePopup/RecipeInfo.cs
@@ -19,7 +19,7 @@ public class RecipeInfo
 
     #region ## Construct ##
 
-    public RecipeInfo(ushort id, RecipeSaveInfo info) : base()
+    public RecipeInfo(int id, RecipeSaveInfo info) : base()
     {
         var data = DataManager.Instance.GetData<RecipeData>();
